[FIX] Groupless backorders list all sales
When a backorder is created from a picking that goes to a customer, a message is added that lists the same sale orders as source as the original picking. However, if the picking in question was not created from a sale, the resulting group will be `False`, and so the search for matching sales will find all sales who do not have a procurement group associated. This is basically any unconfirmed sale in the system at that time.